Refactor:modify the path to worldClass
