+++ src/__tests__/wordSuggestions.test.ts
@@ -6,7 +6,7 @@ import {
   times,
 } from 'lodash';
 import { v4 as uuid } from 'uuid';
-import WordClass from 'src/shared/constants/WordClass';
+import WordClass from 'src/backend/shared/constants/WordClass';
 import Tense from 'src/backend/shared/constants/Tense';
 import SuggestionSource from 'src/backend/shared/constants/SuggestionSource';
 import Dialects from 'src/backend/shared/constants/Dialects';
