+++ src/shared/components/views/shows/WordShow/__tests__/WordShow.test.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { render, configure } from '@testing-library/react';
 import TestContext from 'src/__tests__/components/TestContext';
 import Collections from 'src/shared/constants/Collections';
-import WordClass from 'src/shared/constants/WordClass';
+import WordClass from 'src/backend/shared/constants/WordClass';
 import WordShow from '../WordShow';
 
 configure({ testIdAttribute: 'data-test' });
